[#14] add layout.go to handle layout
