+++ cmd/hashira-cui/layout.go
@@ -1,0 +1,24 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/jroimartin/gocui"
+)
+
+type layout struct {
+}
+
+func (l *layout) Layout(g *gocui.Gui) error {
+	maxX, maxY := g.Size()
+	if v, err := g.SetView("hello", maxX/2-7, maxY/2, maxX/2+7, maxY/2); err != nil {
+		if err != gocui.ErrUnknownView {
+			return err
+		}
+		fmt.Fprintln(v, "Hello world!")
+	}
+	return nil
+}
+func quit(g *gocui.Gui, v *gocui.View) error {
+	return gocui.ErrQuit
+}
